requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
email-validator>=2.2.0
pytest==8.3.3
httpx==0.27.2
